Added Frontend Still Something Needs To Be Done

+++ controllers/product_controller.py
@@ -1,24 +1,34 @@
-from flask import request
+from flask import request,render_template
 from services.products_services import add_items,update_items,delete_item,get_allitems,get_itemsby_type,get_itemby_id
 
 def add_products_controller():
-    data=request.get_json()
-    return add_items(data)
+    if request.method == "POST":
+        data=request.form.to_dict()
+        res= add_items(data)
+        return render_template("index.html",response=res)
+    return render_template("index.html")
 
 def update_products_controller():
-    id=int(request.args.get("id"))
-    data=request.get_json()
-    return update_items(id,data)
+    if request.method == "POST":
+        data=int(request.form.to_dict())
+        return update_items(data)
 
-def delete_products_controller(id):
-    return delete_item(id)
+def delete_products_controller():
+    if request.method == "POST":
+        id=request.form.get("id")
+        return delete_item(id)
 
 def get_all_items_controller():
-    return get_allitems()
+    items= get_allitems()
+    return render_template("index.html",data=items)
 
 def get_itemsby_type_controller():
-    data=request.get_json()
-    return get_itemsby_type(data)
+    type=request.args.get("name")
+    data= get_itemsby_type(type)
+    return render_template("index.html",data=data)
 
-def get_itemby_id_controller(id):
-    return get_itemby_id(id)
+def get_itemby_id_controller():
+    id=request.form.to_dict()
+    items=get_itemby_id(id)
+    data=[items]
+    return render_template("index.html",data=data)
